moved all logic in logic layer

+++ app/src/main/java/com/example/easyshopper/logic/ShoppingListHandler.java
@@ -25,23 +25,41 @@ public class ShoppingListHandler implements Serializable {
 
     //add item into the given shopping list
     public void addItemToList(Product newProduct, ShoppingList shoppingList){
+        if(!shoppingListPersistence.shoppingListExists(shoppingList) ||
+                newProduct == null || shoppingList == null) {
+
+            return;
+        }
+
         shoppingList.addProductToCart(newProduct);
         shoppingListPersistence.updateShoppingList(shoppingList);
     }
 
     //create a new shopping list and add to overall shopping list array
     public void createShoppingList(Store store){
+        //validate inputs
+        if(store == null || shoppingListPersistence.listWithStoreExists(store)) {
+            return;
+        }
+
         ShoppingList newList = new ShoppingList(store);
         shoppingListPersistence.addShoppingList(newList);
     }
 
     //remove given shoppingList from overall shopping list array
     public void removeShoppingList(ShoppingList shoppingList){
-        shoppingListPersistence.deleteShoppingList(shoppingList);
+        if(shoppingList != null && shoppingListPersistence.shoppingListExists(shoppingList)) {
+            shoppingListPersistence.deleteShoppingList(shoppingList);
+        }
     }
 
     //remove an item from a shopping list
     public void removeProduct(Product product, ShoppingList shoppingList){
+        if(!shoppingListPersistence.shoppingListExists(shoppingList)||
+                product == null || shoppingList == null) {
+            return;
+        }
+
         shoppingList.removeProductFromCart(product);
         shoppingListPersistence.updateShoppingList(shoppingList);
     }
@@ -49,6 +67,11 @@ public class ShoppingListHandler implements Serializable {
 
     //Gives total price of the ShoppingList
     public double getCartTotal(ShoppingList shoppingList){
+        if(shoppingList == null || !shoppingListPersistence.shoppingListExists(shoppingList))
+        {
+            return -1;
+        }
+
         double total = 0;
         Store store = shoppingList.getStore();
